いいね非同期化追加完了

+++ app/controllers/public/likes_controller.rb
@@ -8,22 +8,30 @@ class Public::LikesController < ApplicationController
   def create
     post = Post.find(params[:post_id])
     like = current_user.likes.new(post_id: post.id)
+    
     if like.save
       flash[:notice] = "いいねに成功しました。"
+      @post = post
     else
       flash[:alert] = "いいねに失敗しました。"
     end
-    redirect_to post_path(post)
+    render 'replace_btn'
   end
 
   def destroy
     post = Post.find(params[:post_id])
     like = current_user.likes.find_by(post_id: post.id)
+    
+    unless like
+      render 'replace_btn' and return
+    end
+    
     if like.destroy
       flash[:notice] = "いいねを取り消しました。"
+      @post = post
     else
       flash[:alert] = "いいねをの取り消しに失敗しました。"
     end
-    redirect_to post_path(post)
+    render 'replace_btn'
   end
 end
